feat: Add vite_plugin_legacy with tag helpers for @vitejs/plugin-legacy
