+++ examples/rails/app/views/layouts/application.html.erb
@@ -14,7 +14,6 @@
 
   <body>
     <%= yield %>
-    <%= vite_legacy_polyfills_tag %>
     <%= vite_legacy_javascript_tag 'application', asset_type: :typescript %>
   </body>
 </html>
