+++ test/mounted_app/test/dummy/bin/m
@@ -8,14 +8,14 @@
 # this file is here to facilitate running it.
 #
 
-require "pathname"
-ENV["BUNDLE_GEMFILE"] ||= File.expand_path("../../../../../../Gemfile",
+require 'pathname'
+ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../../../../../Gemfile',
   Pathname.new(__FILE__).realpath)
 
-bundle_binstub = File.expand_path("../bundle", __FILE__)
+bundle_binstub = File.expand_path('bundle', __dir__)
 
 if File.file?(bundle_binstub)
-  if File.read(bundle_binstub, 300) =~ /This file was generated by Bundler/
+  if /This file was generated by Bundler/.match?(File.read(bundle_binstub, 300))
     load(bundle_binstub)
   else
     abort("Your `bin/bundle` was not generated by Bundler, so this binstub cannot run.
@@ -23,7 +23,7 @@ Replace `bin/bundle` by running `bundle binstubs bundler --force`, then run this
   end
 end
 
-require "rubygems"
-require "bundler/setup"
+require 'rubygems'
+require 'bundler/setup'
 
-load Gem.bin_path("m", "m")
+load Gem.bin_path('m', 'm')
