+++ vite_plugin_legacy/README.md
@@ -1,0 +1,57 @@
+<h1 align="center">
+  <a href="https://vite-ruby.netlify.app/">
+    <img src="https://raw.githubusercontent.com/ElMassimo/vite_ruby/main/docs/public/logo.svg" width="120px"/>
+  </a>
+
+  <br>
+
+  <a href="https://vite-ruby.netlify.app/">
+    Tag Helpers for <kbd>@vitejs/plugin-legacy</kbd>
+  </a>
+
+  <br>
+
+  <p align="center">
+    <a href="https://github.com/ElMassimo/vite_ruby/actions">
+      <img alt="Build Status" src="https://github.com/ElMassimo/vite_ruby/workflows/build/badge.svg"/>
+    </a>
+    <a href="https://codeclimate.com/github/ElMassimo/vite_ruby">
+      <img alt="Maintainability" src="https://codeclimate.com/github/ElMassimo/vite_ruby/badges/gpa.svg"/>
+    </a>
+    <a href="https://codeclimate.com/github/ElMassimo/vite_ruby">
+      <img alt="Test Coverage" src="https://codeclimate.com/github/ElMassimo/vite_ruby/badges/coverage.svg"/>
+    </a>
+    <a href="https://rubygems.org/gems/vite_plugin_legacy">
+      <img alt="Gem Version" src="https://img.shields.io/gem/v/vite_plugin_legacy.svg?colorB=e9573f"/>
+    </a>
+    <a href="https://github.com/ElMassimo/vite_ruby/blob/master/LICENSE.txt">
+      <img alt="License" src="https://img.shields.io/badge/license-MIT-428F7E.svg"/>
+    </a>
+  </p>
+</h1>
+
+[vite_ruby]: https://github.com/ElMassimo/vite_ruby/tree/main/vite_ruby
+[website]: https://vite-ruby.netlify.app/guide/plugin-legacy
+[plugin-legacy]: https://github.com/vitejs/vite/tree/main/packages/plugin-legacy
+
+__Vite Plugin Legacy__ provides tag helpers to output assets generated by [`@vitejs/plugin-legacy`][plugin-legacy].
+
+## Installation 💿
+
+Add this line to your application's Gemfile:
+
+```ruby
+gem 'vite_plugin_legacy'
+```
+
+Then, run:
+
+```bash
+bundle install
+```
+
+Additional installation instructions are available in the [documentation website][website].
+
+## License
+
+The gem is available as open source under the terms of the [MIT License](https://opensource.org/licenses/MIT).
