+++ vite_plugin_legacy/lib/vite_plugin_legacy.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'vite_ruby'
+
+require 'vite_plugin_legacy/version'
+require 'vite_plugin_legacy/tag_helpers'
+require 'vite_plugin_legacy/engine' if defined?(Rails)
