+++ vite_plugin_legacy/lib/vite_plugin_legacy/engine.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require 'vite_rails'
+require 'rails/railtie'
+
+class VitePluginLegacy::Engine < Rails::Engine
+  initializer 'vite_plugin_legacy.helper' do
+    ActiveSupport.on_load(:action_controller) do
+      ActionController::Base.helper(VitePluginLegacy::TagHelpers)
+    end
+
+    ActiveSupport.on_load(:action_view) do
+      include VitePluginLegacy::TagHelpers
+    end
+  end
+end
