+++ vite_plugin_legacy/lib/vite_plugin_legacy/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module VitePluginLegacy
+  VERSION = '0.5.0'
+end
